requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.6
decorator==5.1.1
docstring-to-markdown==0.12
exceptiongroup==1.1.1
executing==1.2.0
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
-e git+ssh://git@github.com/amodelbello/MWWC_ContactSyncTool.git@3c5666da4cda3c04db7f9d6bc2ddf10a06db4cac#egg=mwwc_sync_contacts
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyairtable==1.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
pytest==7.3.1
pytest-cov==4.1.0
python-dotenv==1.0.0
pytoolconfig==1.2.5
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
ujson==5.7.0
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
